requirements.txt
aiolimiter~=1.0
python-telegram-bot~=20.0
asyncpg~=0.27
greenlet~=2.0
SQLAlchemy~=2.0.1